Add new fields to collection
Add new fields to collection

doesnt break anything just makes more logic to be able to filter your grid based on stock or visibility. FOr instance one could choose to show only visibility=4 so we see the same products as in the frontend

See also here: http://magento.stackexchange.com/questions/17566/magento-add-stock-qty-in-backend-new-order-creation/17898#17898

Optionally we could add the column per default to the grid: although I think this is quite logic and should be in core. Others would maybe not agree. So I will leave that up to you.

        if (Mage::helper('catalog')->isModuleEnabled('Mage_CatalogInventory')) {
            $this->addColumn('inv_qty',
            array(
            'header'=> Mage::helper('catalog')->__('Qty'),
            'width' => '30px',
            'type'  => 'number',
            'index' => 'inv_qty',
            ));
        }