ZendClient breaks when receiving a HTTP/2 response
Ref https://github.com/magento/zf1/pull/23

```
The issue is caused by the fact that the response result is checked by a regular expression that expects the Protocol version in the d.d format (for example, 1.0 or 1.1). For HTTP 2.0, the header returns a value in the format "HTTP/2 200 OK" without the decimal part. This results in an error.
```