Fixes the cloning of a collection by also cloning the internal select object.
Without this fix, the original select object is used and every modification would also affect the original object. A second effect occurs if you clone a collection and use the load function. All the triggers in _beforeLoad function of the collection which add additional filters etc. will be modify the original select object and the select object will be scale up with identical filters and orders.

Signed-off-by: Frank Rochlitzer <f.rochlitzer@b3-it.de>