Throw exception to prevent infinite loop caused by third party code.
Integrate solution from @convenient for issues described in the pull request on magento2 repository
https://github.com/magento/magento2/pull/18678

In our case the problem occurred with the MSP_CashOnDelivery extension, after a long debugging session we were able to identify and reproduce the problem and trace the MSP_CashOnDelivery extension causes an infinite loop and timeout of the request if trigger_recollect column is set to 1 in the sales_flat_quote table

It would be useful to integrate this solution from @convenient that throw new Exception to quickly identify any third party extensions affected by the same problem

But can there be copyright problems in this case? and in the future how can we do for porting a bug fixed on magento 2 to magento-lts?