Only auto-detect the backend_type if none set yet
If an attribute is user defined, the backend type is automatically set every time you save the attribute. If another backend_type was in use before, this breaks the attribute without the admin realizing it. We cannot have a different backend_type unless we also make the attribute is_user_defined = 0. Making the attribute non-user defined, might seam as the solution, but if you do this, the attribute becomes locked in the set and it cannot be removed anymore, so you're stuck with either e locked attribute that must remain in the set, or an attribute that will change backend_type on every save.
My code change prevents the auto-changing of the backend_type. Only if there wasn't a backend_type before, will it now be auto-detected.