Fix newsletter subcriber loading by customer (#1779)
IMHO the implementation of `\Mage_Newsletter_Model_Resource_Subscriber::loadByCustomer` is wrong. Currently, if there are 2 customers on 2 different websites with the same email address, if one customer subscribe himself to the newsletter from one website it will then appear subscribed even from the other website. This is wrong especially if the two websites are synchronized with two different audiences on the mail marketing platform (for example MailChimp). Moreover, if Magento is configured to split customers between websites this implementation could lead to have inconsistent data in the `newsletter_subscriber` table: you could have the subscriber assigned to a `store_id` which is not the the one of its linked customer.

For these reasons I propose this change.