Fixes infinite loop caused by #1467
The infinite loop happens when you try to export collection that has number of items divisible by 1000(`$this->_exportPageSize`).
In this case the Varien collection will reset the page number to `lastPageNumber` when `$page` is higher.
`$page` is higher than `lastPageNumber` because the current logic assume that there is always one more page if the current page count is the same as page export size which is not true in case when count is divisible by page size.