Fix the calculation of bundle saleable status
Mage_Bundle_Model_Product_Type expects the parent::isSaleable() to
return null in the case of a bundle that is not known to be out of
stock, but the parent was flattening that into a boolean false. This
returns the parent to returning null for that case so the specific type
can add its own logic, in the case of bundles taking selection saleable
status into account.