Fixes a notice
After this commit https://github.com/OpenMage/magento-lts/commit/de06e671c09b375605a956e100911396822e276a the code in the checkout has not been updated to look for the session validation data in the new location. As such a PHP notice is generated.

This PR avoids the notice by updating the code to fetch the session validation data in the correct place.