Add null check before method_exists (#2417)
Beginning with version 8.0, PHP throws a type error if the first parameter is neither an object or a class, so the translate method breaks if the translator key is not set to the editor config (which can be the case by the way). We could also use an early return to fix the issue here, but in my opinion the translation method stays understandable and simple enough so the additional check for an existing translator can be added to the IF-statement.