Making optionalZipCountries JS code conditional
`js/varien/form.js` checks for existence of the variable, hence it's allowed to be undefined if empty.

There is more on this. The block is inserted into every Magento page, which seems excessive. It's probably only used in the Cart but I haven't checked thoroughly so leaving it alone for now. Same for the admin.