Add Fedex REST API usage for ship rate and tracking
Fedex is currently in transition from WSDL to REST API.  This change allows configurable usage of WSDL or REST API.

Note: uses shipstream/fedex-rest-sdk, but currently it's not updated with fedex tracking API schema, so need to use a patch branch for it.

Also composer installing shipstream/fedex-rest-sdk requires PHP 8.1+, so I've to change php version as well for this commit.